Cleaned up end2end test for zhmc_partition
Details:
- Moved get_module_output() from utils to test_zhmc_partition.
